build: forge.config.ts
