+++ package.json
@@ -78,11 +78,12 @@
     "@types/debounce": "^1.2.1",
     "@types/debug": "4.1.5",
     "@types/fs-extra": "^9.0.4",
+    "@types/http-proxy": "^1.17.14",
     "@types/jest": "^29.5.2",
     "@types/lodash.isequal": "^4.5.6",
     "@types/lodash.throttle": "^4.1.6",
     "@types/lz-string": "^1.3.34",
-    "@types/node": "^14.14.7",
+    "@types/node": "^20.11.0",
     "@types/query-string": "^6.3.0",
     "@types/react": "^16.9.56",
     "@types/react-addons-css-transition-group": "^15.0.5",
@@ -110,6 +111,7 @@
     "patch-package": "^7.0.0",
     "react-addons-test-utils": "^15.6.2",
     "ts-jest": "^29.1.1",
+    "ts-node": "^10.9.2",
     "typescript": "5.1.6"
   },
   "optionalDependencies": {
